+++ crates/cargo-maelstrom/src/config.rs
@@ -14,6 +14,16 @@ pub struct Config {
 
     #[config(flatten, next_help_heading = "Manifest Config Options")]
     pub cargo_manifest_options: ManifestOptions,
+
+    /// Extra arguments to pass to the test binary. See the help text for a test binary to see what
+    /// it accepts.
+    #[config(
+        var_arg,
+        value_name = "EXTRA-TEST-BINARY-ARGS",
+        default = r#""no args""#,
+        next_help_heading = "Test Binary Options"
+    )]
+    pub extra_test_binary_args: Vec<String>,
 }
 
 impl AsRef<maelstrom_test_runner::config::Config> for Config {