+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `--short` flag added to `maelstrom-go-test`
 - `--fullpath` flag added to `maelstrom-go-test`
 - `extra-test-binary-args` option added to `maelstrom-go-test`
+- `extra-test-binary-args` option added to `cargo-maelstrom`
 
 #### Changed
 - `ui` test-runner configuration option value now matches the same casing as the CLI option